% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DengueBEASTMCC}
\alias{DengueBEASTMCC}
\title{Dengue fever BEAST MCC tree}
\format{A phylo object}
\source{
http://bmcevolbiol.biomedcentral.com/articles/10.1186/1471-2148-7-214
}
\description{
The maximum clade credibility (MCC) tree from \code{\link{DengueTrees}}
}
\references{
Drummond, A. J., and Rambaut, A. (2007) 
BEAST: Bayesian evolutionary analysis by sampling trees.
\emph{BMC Evolutionary Biology}, 7(1), 214.

Lanciotti, R. S., Gubler, D. J., and Trent, D. W. (1997)
Molecular evolution and phylogeny of dengue-4 viruses.
\emph{Journal of General Virology}, 78(9), 2279-2286.
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}
\keyword{datasets}
