% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DengueSeqs}
\alias{DengueSeqs}
\title{Dengue fever sequences}
\format{A DNAbin object containing 17 DNA sequences, each of length 1485.}
\source{
http://bmcevolbiol.biomedcentral.com/articles/10.1186/1471-2148-7-214
}
\description{
17 dengue virus serotype 4 sequences from Lanciotti et al. (1997)
}
\references{
Lanciotti, R. S., Gubler, D. J., and Trent, D. W. (1997)
Molecular evolution and phylogeny of dengue-4 viruses.
\emph{Journal of General Virology}, 78(9), 2279-2286.
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}
\keyword{datasets}
