% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{.render.server.info}
\alias{.render.server.info}
\title{Auxiliary functions}
\usage{
.render.server.info()
}
\description{
These functions are not supposed to be used by the user.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
