% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fluTrees}
\alias{fluTrees}
\title{BEAST analysis of seasonal influenza (A/H3N2)}
\format{A multiPhylo object containing 200 trees, each with 165 tips}
\source{
http://beast.bio.ed.ac.uk/tutorials
}
\description{
These trees were created using BEAST on hemagglutinin (HA) segments
of seasonal influenza A/H3N2 samples collected in New-York city (US)  between 2000 and 2003. This data comes from the influenza BEAST tutorial distributed at:
http://beast.bio.ed.ac.uk/tutorials
}
\details{
Only the first 200 trees (out of 10,000) were retained.
}
\references{
http://beast.bio.ed.ac.uk/tutorials
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\keyword{datasets}
