% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipsMRCAdepths.R
\name{tipsMRCAdepths}
\alias{tipsMRCAdepths}
\title{Tip-tip MRCA depths}
\usage{
tipsMRCAdepths(tree)
}
\arguments{
\item{tree}{An object of class phylo}
}
\description{
This function creates a matrix where columns 1 and 2 correspond to tip labels and column 3 gives the depth of the MRCA of that pair of tips.
It is strongly based on \code{treeVec} and is used by \code{relatedTreeDist} and \code{treeConcordance} where tip labels belong to "categories".
}
\examples{
tree <- rtree(10)
plot(tree)
tipsMRCAdepths(tree)

}
