% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{treespaceServer}
\alias{treespaceServer}
\title{Web-based tree explorer}
\usage{
treespaceServer()
}
\description{
This function opens up an application in a web browser for an interactive exploration of the diversity in a set of trees.
For further details please see the "help" tab within the application.
}
\seealso{
For convenience, \code{treespaceServer} is also available as a separate web app which can be used from any browser (it is not necessary to have \R installed): \url{https://mkendall.shinyapps.io/treespace/}
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}
