% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woodmiceTrees}
\alias{woodmiceTrees}
\title{Bootstrap trees from woodmouse dataset}
\format{A multiPhylo object containing 201 trees, each with 15 tips}
\source{
A set of 15 sequences of the mitochondrial gene cytochrome b of the
woodmouse (Apodemus sylvaticus) which is a subset of the data analysed by
Michaux et al. (2003). The full data set is available through GenBank
(accession numbers AJ511877 to AJ511987)
}
\description{
These trees were created using the neighbour-joining and bootstrapping
example from the ape documentation.
}
\references{
Michaux, J. R., Magnanou, E., Paradis, E., Nieberding, C. and
Libois, R. (2003) Mitochondrial phylogeography of the Woodmouse (Apodemus
sylvaticus) in the Western Palearctic region. \emph{Molecular Ecology}, 12,
685-697
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}
\keyword{datasets}
